# openapi.yaml
openapi: 3.0.3
info:
  title: USBI Platform API
  description: |
    API institucional del sistema USBI. Todos los errores usan RFC 7807 (Problem Details).
    Todos los endpoints autenticados requieren JWT Bearer.
  version: 1.0.0
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: https://api.usbi.edu.mx/api/v1
    description: Servidor de producción (On-Premise, TLS 1.2+)
security:
  - bearerAuth: []
paths:
...
  /sections:
    get:
      summary: Listar secciones oficiales
      operationId: listSections
      parameters:
        - in: query
          name: include_unpublished
          schema:
            type: boolean
            default: false
          description: Sólo admin/operator/director puede ver borradores.
      responses:
        '200':
          description: Lista de secciones visibles para el rol autenticado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SectionListResponseDTO'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      summary: Crear sección oficial
      operationId: createSection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSectionRequestDTO'
      responses:
        '201':
          description: Sección creada como borrador.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SectionDTO'
        '403':
          $ref: '#/components/responses/Forbidden'
  /sections/{section_id}:
    patch:
      summary: Actualizar sección oficial
      operationId: updateSection
      parameters:
        - $ref: '#/components/parameters/SectionID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSectionRequestDTO'
      responses:
        '200':
          description: Sección actualizada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SectionDTO'
        '404':
          $ref: '#/components/responses/NotFound'
  /sections/{section_id}/publish:
    post:
      summary: Publicar sección
      operationId: publishSection
      parameters:
        - $ref: '#/components/parameters/SectionID'
      responses:
        '200':
          description: Sección publicada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SectionDTO'
  /sections/{section_id}/unpublish:
    post:
      summary: Despublicar sección
      operationId: unpublishSection
      parameters:
        - $ref: '#/components/parameters/SectionID'
      responses:
        '200':
          description: Sección regresada a borrador.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SectionDTO'
  /sections/{section_id}/archive:
    post:
      summary: Archivar sección
      operationId: archiveSection
      parameters:
        - $ref: '#/components/parameters/SectionID'
      responses:
        '200':
          description: Sección archivada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SectionDTO'
  /levels:
    get:
      summary: Listar niveles oficiales
      operationId: listLevels
      parameters:
        - in: query
          name: section_id
          schema:
            type: string
            format: uuid
        - in: query
          name: cursor
          schema:
            type: string
            format: uuid
        - in: query
          name: page_size
          schema:
            type: integer
            default: 20
            maximum: 50
        - in: query
          name: include_unpublished
          schema:
            type: boolean
            default: false
          description: Sólo admin/operator/director puede ver borradores.
      responses:
        '200':
          description: Lista paginada de niveles publicados.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LevelListResponseDTO'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      summary: Crear nivel oficial (sólo admin/operator/director)
      operationId: createLevel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLevelRequestDTO'
      responses:
        '201':
          description: Nivel creado en PostgreSQL.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LevelDTO'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /levels/{level_id}:
    get:
      summary: Obtener nivel oficial con contenido
      operationId: getLevel
      parameters:
        - $ref: '#/components/parameters/LevelID'
      responses:
        '200':
          description: Nivel oficial.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LevelDTO'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      summary: Actualizar nivel oficial
      operationId: updateLevel
      parameters:
        - $ref: '#/components/parameters/LevelID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateLevelRequestDTO'
      responses:
        '200':
          description: Nivel actualizado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LevelDTO'
  /levels/{level_id}/publish:
    post:
      summary: Publicar nivel oficial
      operationId: publishLevel
      parameters:
        - $ref: '#/components/parameters/LevelID'
      responses:
        '200':
          description: Nivel publicado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LevelDTO'
  /levels/{level_id}/unpublish:
    post:
      summary: Despublicar nivel oficial
      operationId: unpublishLevel
      parameters:
        - $ref: '#/components/parameters/LevelID'
      responses:
        '200':
          description: Nivel regresado a borrador.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LevelDTO'
  /levels/{level_id}/archive:
    post:
      summary: Archivar nivel oficial
      operationId: archiveLevel
      parameters:
        - $ref: '#/components/parameters/LevelID'
      responses:
        '200':
          description: Nivel archivado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LevelDTO'
  /levels/{level_id}/complete:
    post:
      summary: Completar partida oficial online
      operationId: completeLevel
      parameters:
        - $ref: '#/components/parameters/LevelID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompleteLevelRequestDTO'
      responses:
        '200':
          description: Resultado oficial con XP calculada por backend.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompleteLevelResponseDTO'
...
components:
  parameters:
    SectionID:
      in: path
      name: section_id
      required: true
      schema:
        type: string
        format: uuid
    LevelID:
      in: path
      name: level_id
      required: true
      schema:
        type: string
        format: uuid
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT firmado con HS256. Incluye user_id, role y token_version.
  schemas:
...
    SectionDTO:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        color:
          type: string
        is_published:
          type: boolean
        created_by_admin_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
    CreateSectionRequestDTO:
      type: object
      required: [title, color]
      additionalProperties: false
      properties:
        title:
          type: string
        description:
          type: string
        color:
          type: string
    SectionListResponseDTO:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/SectionDTO'
    LevelDTO:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
          format: uuid
        section_id:
          type: string
          format: uuid
        title:
          type: string
        color:
          type: string
        template_type:
          type: string
          enum: [trivia, puzzle, word_search, fake_news, crossword, memory, snakes_ladders]
        content:
          $ref: '#/components/schemas/LevelContentDTO'
        difficulty:
          type: integer
          minimum: 1
          maximum: 10
        is_published:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    LevelSummaryDTO:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
          format: uuid
        section_id:
          type: string
          format: uuid
        title:
          type: string
        color:
          type: string
        template_type:
          type: string
          enum: [trivia, puzzle, word_search, fake_news, crossword, memory, snakes_ladders]
        difficulty:
          type: integer
          minimum: 1
          maximum: 10
        is_published:
          type: boolean
        created_at:
          type: string
          format: date-time
    CreateLevelRequestDTO:
      type: object
      required: [section_id, title, color, template_type, content, difficulty]
      additionalProperties: false
      properties:
        section_id:
          type: string
          format: uuid
        title:
          type: string
        color:
          type: string
        template_type:
          type: string
          enum: [trivia, puzzle, word_search, fake_news, crossword, memory, snakes_ladders]
        content:
          $ref: '#/components/schemas/LevelContentDTO'
        difficulty:
          type: integer
          minimum: 1
          maximum: 10
    UpdateLevelRequestDTO:
      type: object
      required: [title, color, template_type, content, difficulty]
      additionalProperties: false
      properties:
        title:
          type: string
        color:
          type: string
        template_type:
          type: string
          enum: [trivia, puzzle, word_search, fake_news, crossword, memory, snakes_ladders]
        content:
          $ref: '#/components/schemas/LevelContentDTO'
        difficulty:
          type: integer
          minimum: 1
          maximum: 10
    LevelListResponseDTO:
      type: object
      additionalProperties: false
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/LevelSummaryDTO'
        next_cursor:
          type: string
          format: uuid
    LevelContentDTO:
      oneOf:
        - type: array
          items:
            $ref: '#/components/schemas/TriviaQuestionDTO'
        - $ref: '#/components/schemas/TriviaEnvelopeDTO'
        - $ref: '#/components/schemas/MemoryContentDTO'
        - $ref: '#/components/schemas/FakeNewsContentDTO'
        - $ref: '#/components/schemas/WordSearchContentDTO'
        - $ref: '#/components/schemas/PuzzleContentDTO'
        - $ref: '#/components/schemas/CrosswordContentDTO'
        - $ref: '#/components/schemas/SnakesLaddersContentDTO'
      description: Contenido JSON validado por `template_type`. Máximo 5MB.
    TriviaQuestionDTO:
      type: object
      required: [question, options, correct_index]
      additionalProperties: false
      properties:
        question:
          type: string
          minLength: 1
        options:
          type: array
          minItems: 2
          maxItems: 4
          items:
            type: string
            minLength: 1
        correct_index:
          type: integer
          minimum: 0
          maximum: 3
        media_url:
          type: string
    TriviaEnvelopeDTO:
      type: object
      required: [questions]
      additionalProperties: false
      properties:
        questions:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/TriviaQuestionDTO'
    MemoryPairDTO:
      type: object
      required: [id, content1, content2]
      additionalProperties: false
      properties:
        id:
          type: string
        content1:
          type: string
          minLength: 1
        content2:
          type: string
          minLength: 1
        color:
          type: string
    MemoryContentDTO:
      type: object
      required: [pairs]
      additionalProperties: false
      properties:
        back_color:
          type: string
        pairs:
          type: array
          minItems: 4
          items:
            $ref: '#/components/schemas/MemoryPairDTO'
    FakeNewsItemDTO:
      type: object
      required: [title, content, isFake, reference]
      additionalProperties: false
      properties:
        title:
          type: string
          minLength: 1
        content:
          type: string
          minLength: 1
        isFake:
          type: boolean
        explanation:
          type: string
        imageUrl:
          type: string
        reference:
          type: string
          minLength: 1
    FakeNewsContentDTO:
      type: object
      required: [news]
      additionalProperties: false
      properties:
        news:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/FakeNewsItemDTO'
    WordSearchContentDTO:
      type: object
      required: [words]
      additionalProperties: false
      properties:
        words:
          type: array
          minItems: 2
          items:
            type: string
            minLength: 2
        width:
          type: integer
          minimum: 5
          maximum: 24
        height:
          type: integer
          minimum: 5
          maximum: 24
        seed:
          type: integer
    PuzzleContentDTO:
      type: object
      required: [phrase]
      additionalProperties: false
      properties:
        phrase:
          type: string
          minLength: 1
        pieces:
          type: integer
          minimum: 3
          maximum: 20
          default: 3
        seed:
          type: integer
    CrosswordWordDTO:
      type: object
      required: [word, clue]
      additionalProperties: false
      properties:
        word:
          type: string
          minLength: 2
        clue:
          type: string
          minLength: 1
    CrosswordContentDTO:
      type: object
      required: [words]
      additionalProperties: false
      properties:
        words:
          type: array
          minItems: 2
          description: Las respuestas deben compartir letras suficientes para formar un crucigrama conectado.
          items:
            $ref: '#/components/schemas/CrosswordWordDTO'
    BoardLinkDTO:
      type: object
      required: [start, end]
      additionalProperties: false
      properties:
        start:
          type: integer
          minimum: 1
        end:
          type: integer
          minimum: 1
    SnakesLaddersContentDTO:
      type: object
      required: [board_width, board_height, start_position, end_position]
      additionalProperties: false
      properties:
        board_width:
          type: integer
          minimum: 1
        board_height:
          type: integer
          minimum: 1
        start_position:
          type: integer
          minimum: 1
        end_position:
          type: integer
          minimum: 1
        seed:
          type: integer
        ai_config:
          $ref: '#/components/schemas/SnakesAIConfigDTO'
        questions:
          type: array
          minItems: 3
          items:
            $ref: '#/components/schemas/SnakesQuestionDTO'
        snakes:
          type: array
          items:
            $ref: '#/components/schemas/BoardLinkDTO'
        ladders:
          type: array
          items:
            $ref: '#/components/schemas/BoardLinkDTO'
    SnakesAIConfigDTO:
      type: object
      required: [difficulty]
      additionalProperties: false
      properties:
        difficulty:
          type: string
          enum: [EASY, MEDIUM, HARD]
        fail_probability:
          type: number
          minimum: 0
          maximum: 1
        weights:
          type: array
          items:
            type: number
            minimum: 0
    SnakesQuestionDTO:
      type: object
      required: [question, options, correct_index]
      additionalProperties: false
      properties:
        question:
          type: string
          minLength: 1
        options:
          type: array
          minItems: 2
          maxItems: 4
          items:
            type: string
            minLength: 1
        correct_index:
          type: integer
          minimum: 0
          maximum: 3
    CompleteLevelRequestDTO:
      type: object
      required: [score, completed]
      additionalProperties: false
      properties:
        score:
          type: integer
          minimum: 0
        completed:
          type: boolean
        answers:
          description: Resumen técnico de respuestas. No se usa para calcular XP.
        client_finished_at:
          type: string
          format: date-time
    CompleteLevelResponseDTO:
      type: object
      properties:
        level_id:
          type: string
          format: uuid
        completed:
          type: boolean
        attempt_number:
          type: integer
        xp_awarded:
          type: integer
        total_xp:
          type: integer
        current_streak:
          type: integer
        badges_awarded:
          type: array
          items:
            $ref: '#/components/schemas/BadgeDTO'
    BadgeDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        xp_threshold:
          type: integer
        icon_key:
          type: string
        earned_at:
          type: string
          format: date-time
...
    ProblemDetails:
      type: object
      description: RFC 7807 Problem Details for HTTP APIs
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
          example: https://api.usbi.edu.mx/errors/validation-error
        title:
          type: string
          example: Validation Error
        status:
          type: integer
          example: 400
        detail:
          type: string
          example: El campo email es requerido.
        instance:
          type: string
          format: uri
          example: /api/v1/auth/login
  responses:
    BadRequest:
      description: Solicitud inválida
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    Unauthorized:
      description: No autenticado o JWT expirado/invalidado
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    Forbidden:
      description: Sin permisos para este recurso
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    NotFound:
      description: Recurso no encontrado
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
...
    InternalServerError:
      description: Error interno del servidor
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'